pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dark_subnet"
version = "0.1.0"
description = "Privacy-First Bittensor Subnet with Fully Homomorphic Encryption"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dark Subnet Team"}
]
keywords = ["bittensor", "fhe", "privacy", "machine-learning", "subnet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "bittensor>=6.0.0",
    "concrete-ml>=1.5.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.scripts]
dark-miner = "neurons.miner:main"
dark-validator = "neurons.validator:main"
dark-demo = "demo:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["neurons*", "protocol*", "fhe_models*", "client*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100
